requirements-dev.txt
# pyCGM2@git+https://github.com/qualisys/pyCGM2.git@31d08c9075c8fd2a99a46774a1ee6c8e95133542
requests
pytest
pytest-cov
pytest-xdist
pyinstaller
black
tqdm
beautifulsoup4
PyYAML
xmltodict
ffmpy
# appdirs==1.4.4
# atomicwrites==1.4.0
# attrs==20.3.0
# beautifulsoup4==4.9.3
# black==19.10b0
# certifi
# chardet==4.0.0
# click==7.1.2
# colorama==0.4.4
# configparser==5.0.1
# cycler==0.10.0
# et-xmlfile==1.0.1
# ffmpy==0.3.0
# idna==2.10
# importlib-metadata==3.7.0
# iniconfig==1.1.1
# isort==5.7.0
# jdcal==1.4.1
# keyring==21.1.0
# kiwisolver==1.3.1
# lxml>=4.6.1
# matplotlib==3.3.3
# numpy
# openpyxl==3.0.5
# packaging==20.9
# pandas==1.1.4
# pathspec==0.8.1
# Pillow==8.3.2
# pluggy==0.13.1
# py==1.10.0
# pyparsing==2.4.7
# pytest==6.1.2
# pytest-mpl==0.12
# python-dateutil==2.8.1
# pytest-cov==2.11.1
# pytest-xdist==2.2.1
# psutil==5.8.0
# pytz==2021.1
# pywin32-ctypes==0.2.0
# PyYAML>=5.3.1
# readme-renderer==24.0
# regex==2020.11.13
# requests==2.25.1
# requests-toolbelt==0.9.1
# scipy==1.5.4
# six==1.15.0
# soupsieve==2.2
# toml==0.10.2
# tqdm==4.40.0
# twine==3.1.1
# typed-ast==1.4.2
# typing-extensions==3.7.4.3
# urllib3==1.26.5
# wincertstore==0.2
# xlrd==1.2.0
# xlwt==1.3.0
# xmltodict==0.12.0
# yamlordereddictloader==0.4.0
# zipp==3.4.0
# pyinstaller==5.8.0
